Add some more schedule logging.

+++ xLights/UtilFunctions.h
@@ -5,7 +5,7 @@
 #include <string>
 #include <wx/socket.h>
 
-#define FORMATTIME(ms) (const char *)wxString::Format("%d:%02d.%03d", (ms) / 60000, ((ms) % 60000) / 1000, (ms) % 1000).c_str()
+#define FORMATTIME(ms) (const char *)wxString::Format("%d:%02d.%03d", ((long)ms) / 60000, (((long)ms) % 60000) / 1000, ((long)ms) % 1000).c_str()
 
 // Consolidated set of utility functions
 std::string DecodeIPError(wxSocketError err);
